make quickSearch configurable, refactoring related #89
