--- plugin/src/main/groovy/gorm/tools/dao/MangoDaoEntity.groovy
@@ -1,11 +1,0 @@
-package gorm.tools.dao
-
-import groovy.transform.CompileStatic
-import javax.persistence.Transient
-
-@CompileStatic
-trait MangoDaoEntity {
-
-    @Transient
-    static List<String> quickSearchFields = []
-}
